Migrate reusable qconv bits to helper file
