+++ src/uTensor/CMakeLists.txt
@@ -31,6 +31,7 @@ set(src_utensor_sq_ops
     ops/symmetric_quantization/depthwise_separable_convolution_kernels.cpp
     ops/symmetric_quantization/symmetric_quantization_utils.cpp
     ops/symmetric_quantization/gemmlowp.cpp
+    ops/symmetric_quantization/convolution_helper.cpp
    )
 
 if(X86_PROJECT)
