+++ src/uTensor/ops/symmetric_quantization/depthwise_separable_convolution_kernels.hpp
@@ -5,19 +5,14 @@
 #include "tensor.hpp"
 #include "types.hpp"
 #include "uTensor_util.hpp"
+#include "convolution_helper.hpp"
 
 namespace uTensor {
 namespace TFLM {
 
-DECLARE_ERROR(qDwsMismatchedDimensionsError);
 DECLARE_ERROR(InvalidQDwsActivationRangeError);
 DECLARE_ERROR(InvalidQDwsOutputDepthError);
 
-typedef enum {
-  kTfLitePaddingUnknown = 0,
-  kTfLitePaddingSame,
-  kTfLitePaddingValid,
-} TfLitePadding;
 
 struct TfLiteDepthwiseConvParams {
   // Parameters for DepthwiseConv version 1 or above.
@@ -41,18 +36,6 @@ struct TfLiteDepthwiseConvParams {
   int dilation_height_factor;
 };
 
-enum PaddingType : uint8_t { kNone = 0, kSame, kValid };
-
-struct PaddingValues {
-  int16_t width;
-  int16_t height;
-  // offset is used for calculating "remaining" padding, for example, `width`
-  // is 1 and `width_offset` is 1, so padding_left is 1 while padding_right is
-  // 1 + 1 = 2.
-  int16_t width_offset;
-  // Same as width_offset except it's over the height dimension.
-  int16_t height_offset;
-};
 struct DepthwiseParams {
   PaddingType padding_type;
   PaddingValues padding_values;
@@ -77,55 +60,6 @@ struct DepthwiseParams {
   const int32_t* output_multiplier_per_channel;
   const int32_t* output_shift_per_channel;
 };
-
-struct TfLitePaddingValues {
-  int32_t width;
-  int32_t height;
-  int width_offset;
-  int height_offset;
-};
-
-// It's not guaranteed that padding is symmetric. It's important to keep
-// offset for algorithms need all paddings.
-inline int ComputePaddingWithOffset(int stride, int dilation_rate, int in_size,
-                                    int filter_size, int out_size,
-                                    int* offset) {
-  int effective_filter_size = (filter_size - 1) * dilation_rate + 1;
-  int total_padding =
-      ((out_size - 1) * stride + effective_filter_size - in_size);
-  total_padding = total_padding > 0 ? total_padding : 0;
-  *offset = total_padding % 2;
-  return total_padding / 2;
-}
-
-// Matching GetWindowedOutputSize in TensorFlow.
-inline int ComputeOutSize(TfLitePadding padding, int image_size,
-                          int filter_size, int stride, int dilation_rate = 1) {
-  int effective_filter_size = (filter_size - 1) * dilation_rate + 1;
-  switch (padding) {
-    case kTfLitePaddingSame:
-      return (image_size + stride - 1) / stride;
-    case kTfLitePaddingValid:
-      return (image_size + stride - effective_filter_size) / stride;
-    default:
-      return 0;
-  }
-}
-
-TfLitePaddingValues ComputePaddingHeightWidth(
-    int stride_height, int stride_width, int dilation_rate_height,
-    int dilation_rate_width, int in_height, int in_width, int filter_height,
-    int filter_width, TfLitePadding padding, int* out_height, int* out_width);
-
-void ComputePaddingHeightWidth(int stride_height, int stride_width,
-                               int dilation_rate_height,
-                               int dilation_rate_width, int in_height,
-                               int in_width, int filter_height,
-                               int filter_width, int32_t* padding_height,
-                               int32_t* padding_width, TfLitePadding padding,
-                               int* out_height, int* out_width);
-
-uint16_t MatchingDim(TensorShape s0, uint8_t i0, TensorShape s1, uint8_t i1);
 
 template <typename Tout>
 void DepthwiseConvPerChannel(const DepthwiseParams& params,
